+++ src/main/resources/clcode/default/headers/render.h
@@ -5,16 +5,16 @@
 
 /** Make sure that direction is always normalized!*/
 struct ray_t {
-    double3 origin;
-    double3 direction;
+    float3 origin;
+    float3 direction;
 };
 
 //The parameter for the oneStepSDF function
 struct oneStepSDFArgs_t {
-    double3 point;
+    float3 point;
     __global struct shape_t* shape;
-    double d1;
-    double d2;
+    float d1;
+    float d2;
     int status;
 };
 
@@ -22,19 +22,19 @@ struct oneStepSDFArgs_t {
 struct intersection_t {
     __global struct shape_t* obj;
     struct ray_t* ray;
-    double3 point;
-    double3 normal;
-    double d;
+    float3 point;
+    float3 normal;
+    float d;
 };
 
 __kernel void render(  __write_only image2d_t resultImage,
-                       double4 cameraPosition,
-                       double4 cameraRotation,
-                       double cameraFOV,
+                       float4 cameraPosition,
+                       float4 cameraRotation,
+                       float cameraFOV,
                        int globalNumShapes,
                        __global struct shape_t * globalShapes);
 
-struct ray_t getRay(double u, double v, double3 camPos, double3 camRot, double camFOV);
+struct ray_t getRay(float u, float v, float3 camPos, float3 camRot, float camFOV);
 
 void traceRay(struct ray_t* ray, int numShapes,
             __global struct shape_t* allShapes, struct intersection_t* inter);
@@ -48,25 +48,25 @@ bool firstIntersectionWithPlane(struct ray_t* ray, __global struct shape_t* shap
 
 bool firstIntersectionWithSDF(struct ray_t* ray, __global struct shape_t* shape, struct intersection_t * intersection);
 
-double torusSDF(double3 point, __global struct torusSDF_t* torus);
+float torusSDF(float3 point, __global struct torusSDF_t* torus);
 
-double boxSDF(double3 point, __global struct boxSDF_t* box);
+float boxSDF(float3 point, __global struct boxSDF_t* box);
 
-double subtractionSDF(double3 point, __global struct subtractionSDF_t* subtraction);
+float subtractionSDF(float3 point, __global struct subtractionSDF_t* subtraction);
 
-double distToRay(double3 point, struct ray_t* ray);
+float distToRay(float3 point, struct ray_t* ray);
 
-double distToOrig(struct ray_t* ray);
+float distToOrig(struct ray_t* ray);
 
-double oneStepSDF(double3 point, __global struct shape_t* shape);
+float oneStepSDF(float3 point, __global struct shape_t* shape);
 
 #define EPSILON 0.0001
 
 #define sdfNormal(POINT,SDFFUN,OBJ)\
-    (normalize((double3){\
-            SDFFUN(POINT + (double3){EPSILON,0,0}, OBJ) - SDFFUN(POINT - (double3){EPSILON,0,0}, OBJ),\
-            SDFFUN(POINT + (double3){0,EPSILON,0}, OBJ) - SDFFUN(POINT - (double3){0,EPSILON,0}, OBJ),\
-            SDFFUN(POINT + (double3){0,0,EPSILON}, OBJ) - SDFFUN(POINT - (double3){0,0,EPSILON}, OBJ),\
+    (normalize((float3){\
+            SDFFUN(POINT + (float3){EPSILON,0,0}, OBJ) - SDFFUN(POINT - (float3){EPSILON,0,0}, OBJ),\
+            SDFFUN(POINT + (float3){0,EPSILON,0}, OBJ) - SDFFUN(POINT - (float3){0,EPSILON,0}, OBJ),\
+            SDFFUN(POINT + (float3){0,0,EPSILON}, OBJ) - SDFFUN(POINT - (float3){0,0,EPSILON}, OBJ),\
         }))
 
 #endif
